transfered query preprocess to query handling file
